Test.ThreadNet.General

Documentation

data PropGeneralArgs blk

Constructors

PropGeneralArgs

Fields

pgaBlockProperty ∷ blk → Property

test if the block is as expected

For example, it may fail if the block includes transactions that should have expired before/when the block was forged.

pgaCountTxs ∷ blk → Word64

the number of transactions in the block

pgaExpectedBlockRejection ∷ BlockRejection blk → Bool

whether this block rejection was expected

pgaFirstBlockNo ∷ BlockNo

the block number of the first proper block on the chain

At time of writing this comment... For example, this is 1 for Byron tests and 0 for mock tests. The epoch boundary block (EBB) in slot 0 specifies itself as having block number 0, which implies the genesis block is block number 0, and so the first proper block is number 1. For the mock tests, the first proper block is block number 0.

TODO This implies the mock genesis block does not have a number?

pgaFixedMaxForkLength ∷ Maybe NumBlocks

the maximum length of a unique suffix among the final chains

If not provided, it will be crudely estimated. For example, this estimation is known to be incorrect for PBFT; it does not anticipate PBftExceededSignThreshold.

pgaFixedSchedule ∷ Maybe LeaderSchedule

the leader schedule of the nodes

If not provided, it will be recovered from the nodes' Tracer data.

pgaSecurityParam ∷ SecurityParam

pgaTestConfig ∷ TestConfig

prop_general ∷ ∀ blk. (Condense blk, Eq blk, HasHeader blk, RunNode blk) ⇒ PropGeneralArgs blk → TestOutput blk → Property

The properties always required

Includes:

The competitive chains at the end of the simulation respect the expected bound on fork length

The nodes do not leak file handles

Blocks are exchanged without unexpected delays.

The nodes' chains grow without unexpected delays.

No blocks are unduly rejected (see pgaExpectedBlockRejection).

runTestNetwork

Arguments

∷ (RunNode blk, TxGen blk, TracingConstraints blk)

⇒ TestConfig

→ EpochSize
Temporary: until we start testing the hard fork combinator, we just support a single EpochSize. See also comments for tnaEpochInfo.

→ TestConfigBlock blk

→ TestOutput blk

Thin wrapper around runThreadNetwork

TestConfig

data Rekeying blk

Constructors

Rekeying

Fields

rekeyOracle ∷ CoreNodeId → SlotNo → Maybe SlotNo

The first nominal slot after the given slot, assuming the given core node cannot lead.

rekeyUpd ∷ ProtocolInfo blk → EpochNo → opKey → Maybe (TestNodeInitialization blk)

new config and any corresponding delegation certificate transactions

rekeyFreshSKs ∷ Stream opKey

a stream that only repeats itself after an *effectively* *infinite* number of iterations and also never includes an operational key from the genesis configuration

data TestConfig

Constructors

TestConfig

Fields

numCoreNodes ∷ !NumCoreNodes

numSlots ∷ !NumSlots

TODO generate in function of k

nodeJoinPlan ∷ !NodeJoinPlan

nodeRestarts ∷ !NodeRestarts

nodeTopology ∷ !NodeTopology

slotLength ∷ !SlotLength

initSeed ∷ !Seed

data TestConfigBlock blk

Constructors

TestConfigBlock

Fields

forgeEbbEnv ∷ Maybe (ForgeEbbEnv blk)

nodeInfo ∷ CoreNodeId → TestNodeInitialization blk

rekeying ∷ Maybe (Rekeying blk)

truncateNodeJoinPlan ∷ NodeJoinPlan → NumCoreNodes → (NumSlots, NumSlots) → NodeJoinPlan

truncateNodeRestarts ∷ NodeRestarts → NumSlots → NodeRestarts

truncateNodeTopology ∷ NodeTopology → NumCoreNodes → NodeTopology

Block rejections

data BlockRejection blk

Data about a node rejecting a block as invalid

Constructors

BlockRejection

Fields

brBlockHash ∷ !(HeaderHash blk)

brBlockSlot ∷ !SlotNo

brReason ∷ !(ExtValidationError blk)

brRejector ∷ !NodeId

Re-exports

data ForgeEbbEnv blk

How to forge an EBB

Constructors

ForgeEbbEnv

Fields

forgeEBB ∷ TopLevelConfig blk → SlotNo → BlockNo → ChainHash blk → blk

data TestOutput blk

Constructors

TestOutput

Fields

testOutputNodes ∷ Map NodeId (NodeOutput blk)

testOutputTipBlockNos ∷ Map SlotNo (Map NodeId (WithOrigin BlockNo))

plainTestNodeInitialization ∷ ProtocolInfo blk → TestNodeInitialization blk