Test.ThreadNet.Network

Description

Setup network

Documentation

runThreadNetwork ∷ ∀ m blk. (IOLike m, MonadTimer m, RunNode blk, TxGen blk, TracingConstraints blk, HasCallStack) ⇒ ThreadNetworkArgs m blk → m (TestOutput blk)

Setup a network of core nodes, where each joins according to the node join plan and is interconnected according to the node topology

We run for the specified number of blocks, then return the final state of each node.

data ForgeEbbEnv blk

How to forge an EBB

Constructors

ForgeEbbEnv

Fields

forgeEBB ∷ TopLevelConfig blk → SlotNo → BlockNo → ChainHash blk → blk

type RekeyM m blk

Arguments

= CoreNodeId

→ ProtocolInfo blk

→ SlotNo
The slot in which the node is rekeying

→ (SlotNo → m EpochNo)

→ m (TestNodeInitialization blk)
tniProtocolInfo should include new delegation cert/operational key, and tniCrucialTxs should include the new delegation certificate transaction

How to rekey a node with a fresh operational key

data ThreadNetworkArgs m blk

Parameters for the test node net

Constructors

ThreadNetworkArgs

Fields

tnaForgeEbbEnv ∷ Maybe (ForgeEbbEnv blk)

tnaJoinPlan ∷ NodeJoinPlan

tnaNodeInfo ∷ CoreNodeId → TestNodeInitialization blk

tnaNumCoreNodes ∷ NumCoreNodes

tnaNumSlots ∷ NumSlots

tnaRNG ∷ ChaChaDRG

tnaRekeyM ∷ Maybe (RekeyM m blk)

tnaRestarts ∷ NodeRestarts

tnaSlotLength ∷ SlotLength

tnaTopology ∷ NodeTopology

tnaEpochSize ∷ EpochSize

Epoch size

The ThreadNet tests need to know epoch boundaries in order to know when to insert EBBs, when delegation certificates become active, etc. (This is therefore not related to the chunking of the immutable DB.)

This is temporary: once we have proper support for the hard fork combinator, EpochInfo must be derived from the current ledger state.

tnaTxGenExtra ∷ TxGenExtra blk

data TestNodeInitialization blk

Data used when starting/restarting a node

Constructors

TestNodeInitialization

Fields

tniCrucialTxs ∷ [GenTx blk]

these transactions are added immediately and repeatedly (whenever the ledgerTipSlot changes)

tniProtocolInfo ∷ ProtocolInfo blk

plainTestNodeInitialization ∷ ProtocolInfo blk → TestNodeInitialization blk

type TracingConstraints blk = (Show blk, Show (ApplyTxErr blk), Show (Header blk), Show (GenTx blk), Show (GenTxId blk), ShowQuery (Query blk))

Tracers

data MiniProtocolExpectedException

Non-fatal exceptions expected from the threads of a directedEdge

Constructors

MPEEChainSyncClient ChainSyncClientException
see Ouroboros.Consensus.MiniProtocol.ChainSync.Client

NOTE: the second type in ChainSyncClientException denotes the tip. If it does not agree with the consensus client & server, 'Dynamic chain generation' tests will fail, since they will not catch the right exception.

MPEEBlockFetchClient BlockFetchProtocolFailure
see Ouroboros.Network.BlockFetch.Client

MPEEBlockFetchServer BlockFetchServerException
see Ouroboros.Consensus.MiniProtocol.BlockFetch.Server

MPEETxSubmissionClient TxSubmissionProtocolError
see Ouroboros.Network.TxSubmission.Outbound

MPEETxSubmissionServer TxSubmissionProtocolError
see Ouroboros.Network.TxSubmission.Inbound

data MiniProtocolFatalException

Any synchronous exception from a directedEdge that was not handled as a MiniProtocolExpectedException

Constructors

MiniProtocolFatalException

Fields

mpfeType ∷ !TypeRep

Including the type explicitly makes it easier for a human to debug

mpfeExn ∷ !SomeException

mpfeClient ∷ !CoreNodeId

mpfeServer ∷ !CoreNodeId

data MiniProtocolState

Constructors

MiniProtocolDelayed

MiniProtocolRestarting

data TraceMiniProtocolRestart peer

Constructors

TraceMiniProtocolRestart peer peer SlotNo MiniProtocolState MiniProtocolExpectedException
us them when-start-blocking state reason

Test Output

data TestOutput blk

Constructors

TestOutput

Fields

testOutputNodes ∷ Map NodeId (NodeOutput blk)

testOutputTipBlockNos ∷ Map SlotNo (Map NodeId (WithOrigin BlockNo))

data NodeOutput blk

Constructors

NodeOutput

Fields

nodeOutputAdds ∷ Map SlotNo (Set (RealPoint blk, BlockNo))

nodeOutputFinalChain ∷ Chain blk

nodeOutputFinalLedger ∷ LedgerState blk

nodeOutputForges ∷ Map SlotNo blk

nodeOutputInvalids ∷ Map (RealPoint blk) [ExtValidationError blk]

nodeOutputNodeDBs ∷ NodeDBs MockFS

data NodeDBs db

A vector with an element for each database of a node

The db type parameter is instantiated by this module at types for mock filesystems; either the MockFS type or reference cells thereof.

Constructors

NodeDBs

Fields

nodeDBsImm ∷ db

nodeDBsVol ∷ db

nodeDBsLgr ∷ db