Test.ThreadNet.TxGen

Description

Transaction generator for testing

Documentation

class TxGen blk where

Methods

testGenTxs ∷ MonadRandom m ⇒ NumCoreNodes → SlotNo → TopLevelConfig blk → LedgerState blk → m [GenTx blk]

Generate a number of transactions, valid or invalid, that can be submitted to a node's Mempool.

This function will be called to generate transactions in consensus tests.

Note: this function returns a list so that an empty list can be returned in case we are unable to generate transactions for a blk.