Test.ThreadNet.Util.HasCreator

Description

In tests we want to be able to map a block to the core node that produced it.

Documentation

class HasCreator b where

Methods

getCreator ∷ b → CoreNodeId