Test.ThreadNet.Util.NodeRestarts

Documentation

data NodeRestart

Constructors

NodeRekey
restart the node with a fresh operational key and immediately emit a delegation certificate transaction

NodeRestart
restart the node without rekeying

newtype NodeRestarts

Which nodes are scheduled to restart in each slot

INVARIANT no element Map is empty

Constructors

NodeRestarts (Map SlotNo (Map CoreNodeId NodeRestart))

noRestarts ∷ NodeRestarts

genNodeRestarts ∷ NodeJoinPlan → NumSlots → Gen NodeRestarts

Generate a valid NodeRestarts

POSTCONDITION will not restart a node before it joins

POSTCONDITION will not restart a node when it's scheduled to lead *according* *to* *round-robin*

POSTCONDITION will not simultaneously restart all nodes that have previously joined

shrinkNodeRestarts ∷ NodeRestarts → [NodeRestarts]