Test.ThreadNet.Util.NodeTopology

Node Topology

newtype NodeTopology

Which lesser nodes each node connects to

INVARIANT: for each mapping n -> ms, n > m for each m in ms

INVARIANT: only the mapping for n = CoreNodeId 0 is empty

INVARIANT: there is a mapping for each CoreNodeId in the test, 0 .. n - 1

Note that every node is connected to every other but not necessarily directly. In other words, the topology is always a single network of nodes, which is realistic. With other test configuration components, such as network partitions, the network may split (temporarily or not) into separate connected components during the execution, but the base topology is connected.

Constructors

NodeTopology (Map CoreNodeId (Set CoreNodeId))

coreNodeIdNeighbors ∷ HasCallStack ⇒ NodeTopology → CoreNodeId → [CoreNodeId]

The neighbors of this node

edgesNodeTopology ∷ NodeTopology → [(CoreNodeId, CoreNodeId)]

The edges in this topology

genNodeTopology

Arguments

∷ HasCallStack

⇒ NumCoreNodes
PRECONDITION: non-negative

→ Gen NodeTopology

Generate a NodeTopology consistent with the given properties

shrinkNodeTopology ∷ NodeTopology → [NodeTopology]

Shrink a node topology

The new topologies must be usable with the same number of nodes and slots as the old topology

meshNodeTopology

Arguments

∷ NumCoreNodes
PRECONDITION: non-negative

→ NodeTopology

Connect every pair of nodes

minimumDegreeNodeTopology ∷ NodeTopology → Maybe Int

The neighbor count of the node with the fewest neighbors, unless there are zero nodes