Test.Util.Corruption

Documentation

newtype Corruption

Constructors

Corruption Word

detectCorruption

Arguments

∷ Show a

⇒ (a → Encoding)

→ (∀ s. Decoder s (ByteString → a))

→ (a → Bool)
Integrity check that should detect the corruption. Return False when corrupt.

→ a

→ Corruption

→ Property

Serialise a, apply the given corruption, deserialise it, when that fails, the corruption was detected. When deserialising the corrupted bytestring succeeds, pass the deserialised value to the integrity checking function. If that function returns False, the corruption was detected, if it returns True, the corruption was not detected and the test fails.

applyCorruption ∷ Corruption → ByteString → ByteString

Increment (overflow if necessary) the byte at position i in the bytestring, where i = n mod length bs.

If the bytestring is empty, return it unmodified.