Test.Util.FS.Sim.Error

Description

HasFS instance wrapping SimFS that generates errors, suitable for testing error handling.

Simulate Errors monad

runSimErrorFS ∷ (MonadSTM m, MonadThrow m) ⇒ MockFS → Errors → (StrictTVar m Errors → HasFS m HandleMock → m a) → m (a, MockFS)

Runs a computation provided an Errors and an initial MockFS, producing a result and the final state of the filesystem.

mkSimErrorHasFS ∷ ∀ m. (MonadSTM m, MonadThrow m) ⇒ StrictTVar m MockFS → StrictTVar m Errors → HasFS m HandleMock

Introduce possibility of errors

TODO: Lenses would be nice for the setters

withErrors ∷ MonadSTM m ⇒ StrictTVar m Errors → Errors → m a → m a

Execute the next action using the given Errors. After the action is finished, the previous Errors are restored.

Streams

newtype Stream a

A Stream is a possibly infinite stream of Maybe as.

Constructors

Stream

Fields

getStream ∷ [Maybe a]

mkStream ∷ [Maybe a] → Stream a

Create a Stream based on the given possibly infinite list of Maybe as.

null ∷ Stream a → Bool

Return True if the stream is empty.

A stream consisting of only Nothings (even if it is only one) is not considered to be empty.

runStream ∷ Stream a → (Maybe a, Stream a)

Advance the Stream. Return the Maybe a and the remaining Stream.

always ∷ a → Stream a

Make a Stream that always generates the given a.

mkStreamGen ∷ Int → Gen a → Gen (Stream a)

Make a Stream generator based on a a generator.

The generator generates a finite stream of 10 elements, where each element has a chance of being either Nothing or an element generated with the given a generator (wrapped in a Just).

The first argument is the likelihood (as used by frequency) of a Just where Nothing has likelihood 2.

type ErrorStream = Stream FsErrorType

An ErrorStream is a possibly infinite Stream of (Maybe) FsErrorTypes.

Nothing indicates that there is no error.

Each time the ErrorStream is used (see runErrorStream), the first element (Nothing in case the list is empty) is taken from the list and an ErrorStream with the remainder of the list is returned. The first element represents whether an error should be returned or not.

An FsError consists of a number of fields: fsErrorType, a fsErrorPath, etc. Only the first fields is interesting. Therefore, we only generate the FsErrorType. The FsErrorType will be used to construct the actual FsError.

type ErrorStreamGetSome = Stream (Either FsErrorType Partial)

ErrorStream for hGetSome: an error or a partial get.

type ErrorStreamPutSome = Stream (Either (FsErrorType, Maybe PutCorruption) Partial)

ErrorStream for hPutSome: an error and possibly some corruption, or a partial write.

Generating partial reads/writes

newtype Partial

Given a Partial p where p > 0, we do the following to make a call to hGetSome or hPutSome partial:

hGetSome: we subtract p from the number of requested bytes. If that would result in 0 requested bytes or less, we request 1 byte. If the number of requested bytes was already 0, leave it untouched, as we can't simulate a partial read in this case.

hPutSome: we drop the last p bytes from the bytestring. If that would result in an empty bytestring, just take the first byte of the bytestring. If the bytestring was already empty, leave it untouched, as we can't simulate a partial write in this case.

Constructors

Partial Word64

hGetSomePartial ∷ Partial → Word64 → Word64

hPutSomePartial ∷ Partial → ByteString → ByteString

Generating corruption for hPutSome

data PutCorruption

Model possible corruptions that could happen to a hPutSome call.

Constructors

SubstituteWithJunk Blob
The blob to write is substituted with corrupt junk

PartialWrite Partial
Only perform the write partially

corrupt ∷ ByteString → PutCorruption → ByteString

Apply the PutCorruption to the ByteString.

Error streams for HasFS

data Errors

Error streams for the methods of the HasFS type class.

An ErrorStream is provided for each method of the HasFS type class. This ErrorStream will be used to generate potential errors that will be thrown by the corresponding method.

For hPutSome, an ErrorStreamWithCorruption is provided to simulate corruption.

An Errors is used in conjunction with SimErrorFS, which is a layer on top of SimFS that simulates methods throwing FsErrors.

Constructors

Errors

Fields

dumpStateE ∷ ErrorStream

hOpenE ∷ ErrorStream

hCloseE ∷ ErrorStream

hSeekE ∷ ErrorStream

hGetSomeE ∷ ErrorStreamGetSome

hGetSomeAtE ∷ ErrorStreamGetSome

hPutSomeE ∷ ErrorStreamPutSome

hTruncateE ∷ ErrorStream

hGetSizeE ∷ ErrorStream

createDirectoryE ∷ ErrorStream

createDirectoryIfMissingE ∷ ErrorStream

listDirectoryE ∷ ErrorStream

doesDirectoryExistE ∷ ErrorStream

doesFileExistE ∷ ErrorStream

removeFileE ∷ ErrorStream

renameFileE ∷ ErrorStream

genErrors

Arguments

∷ Bool
True -> generate partial writes

→ Bool
True -> generate SubstituteWithJunk corruptions

→ Gen Errors

Generator for Errors that allows some things to be disabled.

This is needed by the VolatileDB state machine tests, which try to predict what should happen based on the Errors, which is too complex sometimes.

allNull ∷ Errors → Bool

Return True if all streams are empty (null).

simpleErrors ∷ ErrorStream → Errors

Use the given ErrorStream for each field/method. No corruption of hPutSome.