Test.Util.Golden

Documentation

goldenTestCBOR

Arguments

∷ HasCallStack

⇒ (a → Encoding)
Encoder

→ a
Example value to use for the golden test

→ FlatTerm
Expected CBOR (flat) encoding

→ Assertion

Test whether the encoding of the given example value matches the expected encoding, given as a FlatTerm.

Hint: when writing the test, just pass an empty expected FlatTerm, a pastable one will be printed.

Orphan instances

Generic TermToken

Instance details

Associated Types

type Rep TermToken ∷ Type → Type

Methods

from ∷ TermToken → Rep TermToken x

to ∷ Rep TermToken x → TermToken

ToExpr TermToken

Instance details

Methods

toExpr ∷ TermToken → Expr

listToExpr ∷ [TermToken] → Expr