Test.Util.LogicalClock

Description

Logical time (in terms of abstract " ticks ")

Intended for qualified import

import Test.Util.LogicalClock (LogicalClock)
import qualified Test.Util.LogicalClock as LogicalClock

API

data LogicalClock m

Logical clock (in terms of ticks rather than actual UTCTime)

Constructors

LogicalClock

Fields

getCurrentTick ∷ STM m Tick

Get the current " time "

waitUntilDone ∷ m ()

Wait for the end of time (each clock has a maximum number of ticks)

mockSystemTime ∷ SystemTime m

Translate the logical clock to mock SystemTime

newtype Tick

Logical time unit

Constructors

Tick

Fields

tickToWord64 ∷ Word64

newtype NumTicks

Number of ticks the test will run for

Constructors

NumTicks Word64

Construction

new ∷ IOLike m ⇒ ResourceRegistry m → NumTicks → m (LogicalClock m)

sufficientTimeFor ∷ HasCallStack ⇒ [Tick] → NumTicks

Set NumTicks such that we will have seen all of the specified Ticks

Scheduling actions

onEachTick ∷ (IOLike m, HasCallStack) ⇒ ResourceRegistry m → LogicalClock m → String → (Tick → m ()) → m (m ())

Execute action on every clock tick

Returns a handle to cancel the thread.

onTick ∷ (IOLike m, HasCallStack) ⇒ ResourceRegistry m → LogicalClock m → String → Tick → m () → m ()

Execute action once at the specified tick

blockUntilTick ∷ MonadSTM m ⇒ LogicalClock m → Tick → m Bool

Block until the specified tick

Returns False if the current tick is later than the requested one, or True if they were equal.

Conversions

hardForkBlockchainTime ∷ (IOLike m, HasHardForkHistory blk, HasCallStack) ⇒ ResourceRegistry m → Tracer m TraceBlockchainTimeEvent → LogicalClock m → LedgerConfig blk → STM m (LedgerState blk) → m (BlockchainTime m)

Construct BlockchainTime using hardForkBlockchainTime

We do this by providing it a SystemTime instance that translates ticks to time.

checkInFuture ∷ (Monad m, HasHardForkHistory blk, UpdateLedger blk) ⇒ LedgerConfig blk → LogicalClock m → CheckInFuture m blk

Only exported for the benefit of WrappedClock

newWithDelay ∷ (IOLike m, HasCallStack) ⇒ ResourceRegistry m → NumTicks → NominalDiffTime → m (LogicalClock m)

Generalization of new that allows to override the tickDelay

NOTE: Tests using the logical clock really should not need to know what the tick delay is; that's kind of the point of a logical clock after all.