Test.Util.Orphans.Arbitrary

Documentation

genLimitedEpochSize ∷ Gen EpochSize

The tests for CumulEpochSizes requires that the sum of a list of these values does not overflow.

An epoch size must be > 0.

genLimitedSlotNo ∷ Gen SlotNo

The functions slotAtTime and timeUntilNextSlot suffer from arithmetic overflow for very large values, so generate values that avoid overflow when used in these two functions. The largest value generated is still sufficently large to allow for 5e12 years worth of slots at a slot interval of 20 seconds.

genSmallEpochNo ∷ Gen EpochNo

genSmallSlotNo ∷ Gen SlotNo

Generate a small SlotNo for the state machine tests. The runtime of the StateMachine prop_sequential tests is proportional the the upper bound.

newtype SmallDiffTime

Wrapper around NominalDiffTime with custom Arbitrary instance

The default Arbitrary instance for NominalDiffTime isn't very useful:

It tends to pick huge values

It tends not to pick integer values

It does not shrink

Our custom instance

Picks values between 0 and (1000 * 20 * 10) seconds:

Maximum segment length: 1000

Maximum slot length: 20 seconds

Maximum number of segments: 10

With a 0.1 second precision

Shrinks

Constructors

SmallDiffTime NominalDiffTime

dawnOfTime ∷ UTCTime

Dawn of time

Everybody knows nothing happened before 2000-01-01 00:00:00

Orphan instances

Arbitrary SlotNo

Instance details

Methods

arbitrary ∷ Gen SlotNo

shrink ∷ SlotNo → [SlotNo]

Arbitrary EpochNo

Instance details

Methods

arbitrary ∷ Gen EpochNo

shrink ∷ EpochNo → [EpochNo]

Arbitrary NominalDiffTime
Picks time span between 0 seconds and (roughly) 50 years

Instance details

Methods

arbitrary ∷ Gen NominalDiffTime

shrink ∷ NominalDiffTime → [NominalDiffTime]

Arbitrary UTCTime
Picks moment between dawnOfTime and (roughly) 50 years later

Uses instance for NominalDiffTime

Instance details

Methods

arbitrary ∷ Gen UTCTime

shrink ∷ UTCTime → [UTCTime]

Arbitrary ClockSkew

Instance details

Methods

arbitrary ∷ Gen ClockSkew

shrink ∷ ClockSkew → [ClockSkew]

Arbitrary ChunkSlot

Instance details

Methods

arbitrary ∷ Gen ChunkSlot

shrink ∷ ChunkSlot → [ChunkSlot]

Arbitrary ChunkSize
Picks a ChunkSize between 1 and 100, and randomly choose to enable EBBs

Instance details

Methods

arbitrary ∷ Gen ChunkSize

shrink ∷ ChunkSize → [ChunkSize]

Arbitrary ChunkNo
This picks an EpochNo between 0 and 10000

We don't pick larger values because we're not interested in testing overflow due to huge epoch numbers and even huger slot numbers.

Instance details

Methods

arbitrary ∷ Gen ChunkNo

shrink ∷ ChunkNo → [ChunkNo]

Arbitrary RelativeSlot

Instance details

Methods

arbitrary ∷ Gen RelativeSlot

shrink ∷ RelativeSlot → [RelativeSlot]

Arbitrary NumCoreNodes

Instance details

Methods

arbitrary ∷ Gen NumCoreNodes

shrink ∷ NumCoreNodes → [NumCoreNodes]

Arbitrary Seed

Instance details

Methods

arbitrary ∷ Gen Seed

shrink ∷ Seed → [Seed]

Arbitrary SystemStart

Instance details

Methods

arbitrary ∷ Gen SystemStart

shrink ∷ SystemStart → [SystemStart]

Arbitrary SlotLength
Length between 0.001 and 20 seconds, millisecond granularity

Instance details

Methods

arbitrary ∷ Gen SlotLength

shrink ∷ SlotLength → [SlotLength]

Arbitrary NumSlots

Instance details

Methods

arbitrary ∷ Gen NumSlots

shrink ∷ NumSlots → [NumSlots]