Test.Util.QuickCheck

Description

QuickCheck utilities

Comparison functions

lt ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like <, but prints a counterexample when it fails.

ge ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like >=, but prints a counterexample when it fails.

expectRight ∷ (Show a, Show b, Eq b) ⇒ b → Either a b → Property

Check that we have the expected Right value

expectRight b ab is roughly equivalent to Right b === ab, but avoids an equality constraint on a.

Improved variants

elements ∷ HasCallStack ⇒ [a] → Gen a

Generates one of the given values. The input list must be non-empty.

NOTE unlike the standard elements, this variant has a HasCallStack constraint, which makes debugging the error much easier.

(=:=) ∷ (Eq a, Condense a) ⇒ a → a → Property infix 4

Like ===, but uses Condense instead of Show when it fails.

Convenience

collects ∷ Show a ⇒ [a] → Property → Property