Test.Util.Random

Documentation

runMonadRandomWithTVar ∷ MonadSTM m ⇒ StrictTVar m ChaChaDRG → RunMonadRandom m

A RunMonadRandom record to use in cases where we don't have access to the system random number generator, e.g., in the tests.

We store a ChaChaDRG in a TVar, and whenever we need to run a computation requiring MonadRandom, we split the ChaChaDRG, store a split half back in the TVar and run the computation using the other half. Each random computation will use a different DRG, even concurrent ones.

Remember that will only be used in the tests code, in the real implementation we will use runMonadRandomIO.