Test.Util.Serialisation

Documentation

roundtrip_all ∷ ∀ blk. (SerialiseDiskConstraints blk, SerialiseNodeToNodeConstraints blk, SerialiseNodeToClientConstraints blk, Show (NodeToNodeVersion blk), Show (NodeToClientVersion blk), HasAnnTip blk, Arbitrary' blk, Arbitrary' (Header blk), Arbitrary' (HeaderHash blk), Arbitrary' (LedgerState blk), Arbitrary' (AnnTip blk), Arbitrary' (ConsensusState (BlockProtocol blk)), ArbitraryWithVersion (NodeToNodeVersion blk) (HeaderHash blk), ArbitraryWithVersion (NodeToNodeVersion blk) blk, ArbitraryWithVersion (NodeToNodeVersion blk) (Header blk), ArbitraryWithVersion (NodeToNodeVersion blk) (GenTx blk), ArbitraryWithVersion (NodeToNodeVersion blk) (GenTxId blk), ArbitraryWithVersion (NodeToNodeVersion blk) (SomeBlock (NestedCtxt Header) blk), ArbitraryWithVersion (NodeToClientVersion blk) (HeaderHash blk), ArbitraryWithVersion (NodeToClientVersion blk) blk, ArbitraryWithVersion (NodeToClientVersion blk) (GenTx blk), ArbitraryWithVersion (NodeToClientVersion blk) (ApplyTxErr blk), ArbitraryWithVersion (NodeToClientVersion blk) (Some (Query blk)), ArbitraryWithVersion (NodeToClientVersion blk) (SomeResult blk)) ⇒ CodecConfig blk → (∀ a. NestedCtxt_ blk Header a → Dict (Eq a, Show a)) → TestTree

roundtrip_SerialiseDisk ∷ ∀ blk. (SerialiseDiskConstraints blk, HasAnnTip blk, Arbitrary' blk, Arbitrary' (Header blk), Arbitrary' (HeaderHash blk), Arbitrary' (LedgerState blk), Arbitrary' (AnnTip blk), Arbitrary' (ConsensusState (BlockProtocol blk))) ⇒ CodecConfig blk → (∀ a. NestedCtxt_ blk Header a → Dict (Eq a, Show a)) → [TestTree]

roundtrip_SerialiseNodeToNode ∷ ∀ blk. (SerialiseNodeToNodeConstraints blk, Show (NodeToNodeVersion blk), ArbitraryWithVersion (NodeToNodeVersion blk) (HeaderHash blk), ArbitraryWithVersion (NodeToNodeVersion blk) blk, ArbitraryWithVersion (NodeToNodeVersion blk) (Header blk), ArbitraryWithVersion (NodeToNodeVersion blk) (GenTx blk), ArbitraryWithVersion (NodeToNodeVersion blk) (GenTxId blk), EncodeDisk blk blk, DecodeDisk blk (ByteString → blk), HasNestedContent Header blk, EncodeDiskDep (NestedCtxt Header) blk, DecodeDiskDep (NestedCtxt Header) blk) ⇒ CodecConfig blk → [TestTree]

roundtrip_SerialiseNodeToClient ∷ ∀ blk. (SerialiseNodeToClientConstraints blk, Show (NodeToClientVersion blk), ArbitraryWithVersion (NodeToClientVersion blk) (HeaderHash blk), ArbitraryWithVersion (NodeToClientVersion blk) blk, ArbitraryWithVersion (NodeToClientVersion blk) (GenTx blk), ArbitraryWithVersion (NodeToClientVersion blk) (ApplyTxErr blk), ArbitraryWithVersion (NodeToClientVersion blk) (Some (Query blk)), ArbitraryWithVersion (NodeToClientVersion blk) (SomeResult blk), EncodeDisk blk blk, DecodeDisk blk (ByteString → blk)) ⇒ CodecConfig blk → [TestTree]

roundtrip_envelopes ∷ ∀ blk. (SerialiseNodeToNode blk (SerialisedHeader blk), HasNestedContent Header blk) ⇒ CodecConfig blk → WithVersion (NodeToNodeVersion blk) (SomeBlock (NestedCtxt Header) blk) → Property

This is similar to the roundtrip tests for headers, except we don't start with a header but some fixed bytestring in the payload. This makes debugging a bit easier as we can focus on just the envelope.

data WithVersion v a

Used to generate arbitrary values for the serialisation roundtrip tests. As the serialisation format can change with the version, not all arbitrary values of the type might be valid for each version.

For example, a certain constructor can only be used after a certain version and can thus not be generated for any prior versions.

Constructors

WithVersion v a

SomeResult

data SomeResult blk where

To easily generate all the possible results of the Query GADT, we introduce an existential that also bundles the corresponding Query as evidence. We also capture Eq, Show, and Typeable constraints, as we need them in the tests.

Constructors

SomeResult ∷ (Eq result, Show result, Typeable result) ⇒ Query blk result → result → SomeResult blk