Test.Util.Shrink

Description

Utility functions for defining QuickCheck's shrink

Documentation

andId ∷ (a → [a]) → a → [a]

Add the argument as the last element of the output

dropId ∷ [a] → [a]

Drop the last element

If every source list in a comprehension uses andId, then the last element will be a copy of the initial input.