Test.Util.Time

Test blockchain time API

data TestBlockchainTime m

Test blockchain time

Unlike the real BlockchainTime API, the TestBlockchainTime API always knows what the current SlotNo is.

In addition, TestBlockchainTime defines a maximum number of slots that the system will run for, and offers an API for waiting until that slot.

TODO: We will eventually need to generalize the tests also so that they do not rely on knowing the time.

testBlockchainTimeSlot ∷ TestBlockchainTime m → STM m SlotNo

testBlockchainTimeDone ∷ TestBlockchainTime m → m ()

Blocks until the end of the final requested slot.

Construction

newtype NumSlots

Number of slots

Constructors

NumSlots Word64

newTestBlockchainTime

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ NumSlots
Number of slots

→ SlotLength
Slot duration

→ m (TestBlockchainTime m)

Construct new blockchain time that ticks at the specified slot duration

NOTE: This is just one way to construct time. We can of course also connect this to the real time (if we are in IO), or indeed to a manual tick (in a demo).

NOTE: The number of slots is only there to make sure we terminate the thread (otherwise the system will keep waiting).

NOTE: Any code not passed to onSlotChange may start running "before" the first slot SlotNo 0, i.e. during Initializing. This is likely only appropriate for initialization code etc. In contrast, the argument to onSlotChange is blocked at least until SlotNo 0 begins.

Alternative constructors

fixedBlockchainTime ∷ MonadSTM m ⇒ SlotNo → TestBlockchainTime m

TestBlockchainTime that is stuck on the given slot

Derived funtionality

testBlockchainTime ∷ TestBlockchainTime m → BlockchainTime m

blockUntilSlot ∷ IOLike m ⇒ TestBlockchainTime m → SlotNo → m Bool

Block until the specified slot

Returns True immediately if the requested slot is already over, else blocks as requested and then returns False

onSlotChange

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ TestBlockchainTime m

→ String
Label for the thread

→ (SlotNo → m ())
Action to execute

→ m (m ())

Variant on onKnownSlotChange

Executing an action once

data OnSlotException

Constructors

OnSlotTooLate SlotNo SlotNo
An action was scheduled via onSlot for a slot in the past. First slot is requested, second slot is current as of raising.

onSlot

Arguments

∷ (HasCallStack, IOLike m)

⇒ ResourceRegistry m

→ TestBlockchainTime m

→ String
Label for the thread

→ SlotNo

→ m ()

→ m ()

Run the specified action once when the specified slot is reached