Test.Util.Tracer

Documentation

recordingTracerIORef ∷ IO (Tracer IO ev, IO [ev])

Create a Tracer that stores all events in an IORef that is atomically updated. The second return value lets you obtain the events recorded so far (from oldest to newest). Obtaining the events does not erase them.

recordingTracerTVar ∷ MonadSTM m ⇒ m (Tracer m ev, m [ev])

Create a Tracer that stores all events in a TVar that is atomically updated. The second return value lets you obtain the events recorded so far (from oldest to newest). Obtaining the events does not erase them.