Test.Util.WithEq

Documentation

data WithEq a

Use this type to add an Eq instance for types that don't have one or for which one doesn't make sense, but an Eq instance is needed for testing purposes.

E.g., Iterator needs an Eq instance in the q-s-m tests

Constructors

WithEq

Fields

getId ∷ Id

unWithEq ∷ a

newtype Id

Constructors

Id Word