Test.Util.WrappedClock

Description

Intended for qualified import

import Test.Util.WrappedClock (WrappedClock(..))
import qualified Test.Util.WrappedClock as WrappedClock

Documentation

data WrappedClock m

Wrapper around LogicalClock that pretends Tick == SlotNo

The only reason this type exists is to ease the transition from our innocent youth in which we thought the current slot was always known to the harsh reality where it turns out the current slot is ledger state dependent. We should endeavour to reduce to the scope of the WrappedClock as much as possible (unwrapping it to use the nested LogicalClock), and eventually remove this type altogether. Indeed, if we want to test our test across the hard fork we will have no choice (unless we pick the same slot length I guess).

Constructors

WrappedClock

Fields

unwrapLogicalClock ∷ LogicalClock m

newtype NumSlots

Number of slots

Constructors

NumSlots Word64

Wrap LogicalClock API

new ∷ (IOLike m, HasCallStack) ⇒ ResourceRegistry m → NumSlots → SlotLength → m (WrappedClock m)

getCurrentSlot ∷ MonadSTM m ⇒ WrappedClock m → STM m SlotNo

waitUntilDone ∷ WrappedClock m → m ()

onSlotChange

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ WrappedClock m

→ String
Label for the thread

→ (SlotNo → m ())
Action to execute

→ m (m ())

onSlot ∷ (HasCallStack, IOLike m) ⇒ ResourceRegistry m → WrappedClock m → String → SlotNo → m () → m ()

blockUntilSlot ∷ IOLike m ⇒ WrappedClock m → SlotNo → m Bool

Conversions

blockchainTime ∷ MonadSTM m ⇒ WrappedClock m → BlockchainTime m