Test.Util.Roundtrip

Documentation

roundtrip ∷ (Eq a, Show a) ⇒ (a → Encoding) → (∀ s. Decoder s a) → a → Property

roundtrip'

Arguments

∷ (Eq a, Show a)

⇒ (a → Encoding)
enc

→ (∀ s. Decoder s (ByteString → a))

→ a

→ Property

Roundtrip property for values annotated with their serialized form

NOTE: Suppose a consists of a pair of the unannotated value a' and some ByteString. The roundtrip property will fail if that ByteString encoding is not equal to enc a'. One way in which this might happen is if the annotation is not canonical CBOR, but enc does produce canonical CBOR.