Test.Util.Time

Documentation

newtype NumSlots

Number of slots

Constructors

NumSlots Word64

data TestBlockchainTime m

testBlockchainTimeSlot ∷ TestBlockchainTime m → STM m SlotNo

testBlockchainTimeDone ∷ TestBlockchainTime m → m ()

Blocks until the end of the final requested slot.

newTestBlockchainTime

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ NumSlots
Number of slots

→ SlotLength
Slot duration

→ m (TestBlockchainTime m)

Construct new blockchain time that ticks at the specified slot duration

NOTE: This is just one way to construct time. We can of course also connect this to the real time (if we are in IO), or indeed to a manual tick (in a demo).

NOTE: The number of slots is only there to make sure we terminate the thread (otherwise the system will keep waiting).

NOTE: Any code not passed to onSlotChange may start running "before" the first slot SlotNo 0, i.e. during Initializing. This is likely only appropriate for initialization code etc. In contrast, the argument to onSlotChange is blocked at least until SlotNo 0 begins.

blockUntilSlot ∷ IOLike m ⇒ TestBlockchainTime m → SlotNo → m Bool

Block until the specified slot

Returns True immediately if the requested slot is already over, else blocks as requested and then returns False

testBlockchainTime ∷ TestBlockchainTime m → BlockchainTime m